ai/requirements.txt
av==12.1.0
fastapi==0.111.0
numpy==1.26.4
Pillow==10.3.0
torch==2.3.0+cu121
transformers==4.41.1
uvicorn==0.30.1
